+++ askbot/tests/form_tests.py
@@ -160,11 +160,7 @@ class EditQuestionAnonymouslyFormTests(AskbotTestCase):
         (0, 1, 1, 0, False),
         (0, 1, 1, 1, False),#all up to this point are False
         (1, 0, 0, 0, False),
-        (1, 0, 0, 1, 'error'),#not owner
-        (1, 0, 1, 0, 'error'),#rules changed either reload page or check box
-        (1, 0, 1, 1, True),#rules changed - say yes here
         (1, 1, 0, 0, False),
-        (1, 1, 0, 1, 'error'),
         (1, 1, 1, 0, False),
         (1, 1, 1, 1, True),
     )
@@ -214,7 +210,7 @@ class EditQuestionAnonymouslyFormTests(AskbotTestCase):
             self.setup_data(*(entry[:4]))
 
             if self.form.is_valid():
-                result = self.form.cleaned_data['reveal_identity']
+                result = self.form.cleaned_data.get('reveal_identity', False)
             else:
                 result = 'error'